refactor: isolate admin audit sink
