+++ crates/dcc-mcp-gateway/src/gateway/admin/infra/mod.rs
@@ -3,11 +3,13 @@
 //! Contains modules that perform I/O and data aggregation against external stores.
 
 pub mod activity;
+pub(crate) mod audit_sink;
 
 #[cfg(feature = "admin")]
 pub(crate) mod integrations;
 
 pub use activity::*;
+pub use audit_sink::AdminAuditSink;
 
 #[cfg(feature = "admin")]
 pub use integrations::*;
